fixes runBuilder and runMutator scripts from T12

+++ T12_mull_gtest_cmake_split/runBuilder.sh
@@ -1,5 +1,12 @@
-rm -rf ./build/*
-export CXX=clang++-12
+#!/bin/bash
+
+if [ -d 'build' ] ; then
+    rm -rf ./build/*
+    echo "tem build"
+else
+    mkdir build
+fi
 cd build
+export CXX=clang++-12
 cmake -DCMAKE_CXX_FLAGS="-O0 -fexperimental-new-pass-manager -fpass-plugin=/usr/lib/mull-ir-frontend-12 -g -grecord-command-line" ..
 make  -j
